CommentableContainer refact and Page renders
Functions not clearly working.  Got rid of comm_it/type in favor a
commentable object (movie or comment)

To reduce confusion.  Comments are owned by commentables.

So using the correct id was confusing

Need to see how this flows thorugh the app.  TDD is always worth it in the end

Wish I'd done that!